Notification Detail Mobile - mooved common components
