--- packages/pn-pa-webapp/src/pages/components/NotificationDetail/__test__/DetailDocuments.test.tsx
@@ -1,47 +1,0 @@
-import { fireEvent, waitFor, RenderResult } from '@testing-library/react';
-import * as redux from 'react-redux';
-
-import { render } from '../../../../__test__/test-utils';
-import * as actions from '../../../../redux/notification/actions';
-import { notificationToFe } from '../../../../redux/notification/__test__/test-utils';
-import DetailDocuments from '../DetailDocuments';
-
-describe('Notification Detail Documents Component', () => {
-  let result: RenderResult | undefined;
-  let mockDispatchFn: jest.Mock;
-
-  beforeEach(() => {
-    // mock dispatch
-    const useDispatchSpy = jest.spyOn(redux, 'useDispatch'); 
-    mockDispatchFn = jest.fn();
-    useDispatchSpy.mockReturnValue(mockDispatchFn);
-    // render component
-    result = render(<DetailDocuments notification={notificationToFe}/>);
-  });
-
-  afterEach(() => {
-    result = undefined;
-    jest.clearAllMocks();
-  });
-
-  it('renders detail documents', () => {
-    expect(result?.container).toHaveTextContent(/Atti Allegati/i);
-    // expect(result?.container).toHaveTextContent(/Scarica tutti gli Atti/i);
-    const documentsButtons = result?.getAllByTestId('documentButton');
-    expect(documentsButtons).toHaveLength(notificationToFe.documents.length);
-  });
-
-  it('test click on document button', async () => {
-    const documentsButtons = result?.getAllByTestId('documentButton');
-    const actionSpy = jest.spyOn(actions, 'getSentNotificationDocument');
-    const mockActionFn = jest.fn();
-    actionSpy.mockImplementation(mockActionFn);
-    await waitFor(() => {
-      fireEvent.click(documentsButtons![0]);
-    });
-    expect(mockDispatchFn).toBeCalledTimes(1);
-    expect(mockActionFn).toBeCalledTimes(1);
-    expect(mockActionFn).toBeCalledWith({iun: notificationToFe.iun, documentIndex: 0});
-  });
-
-});
