+++ packages/pn-personafisica-webapp/src/pages/NotificationDetail.page.tsx
@@ -1,17 +1,21 @@
 import { useParams, useNavigate } from 'react-router-dom';
-import { useEffect } from 'react';
+import { ReactNode, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Breadcrumbs, Grid, Typography, Box, Paper, Button } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import EmailIcon from '@mui/icons-material/Email';
-import { TitleBox } from '@pagopa-pn/pn-commons';
+import { TitleBox, LegalFactId,
+  NotificationDetailDocuments,
+  NotificationDetailTable,
+  NotificationDetailTimeline, } from '@pagopa-pn/pn-commons';
+
 import * as routes from '../navigation/routes.const';
 import { useAppDispatch, useAppSelector } from '../redux/hooks';
 import { RootState } from '../redux/store';
-import DetailTimeline from '../component/NotificationDetail/DetailTimeline';
-import DetailTable from '../component/NotificationDetail/DetailTable';
-import DetailDocuments from '../component/NotificationDetail/DetailDocuments';
-import { getReceivedNotification, resetState } from '../redux/notification/actions';
+import { getReceivedNotification,
+  getReceivedNotificationDocument,
+  getReceivedNotificationLegalfact,
+  resetState, } from '../redux/notification/actions';
 import StyledLink from '../component/StyledLink/StyledLink';
 
 const useStyles = makeStyles(() => ({
@@ -29,12 +33,67 @@ const NotificationDetail = () => {
   const notification = useAppSelector((state: RootState) => state.notificationState.notification);
   const navigate = useNavigate();
   const { t } = useTranslation(['notifiche', 'common']);
+  const documentDownloadUrl = useAppSelector(
+    (state: RootState) => state.notificationState.documentDownloadUrl
+  );
+  const legalFactDownloadUrl = useAppSelector(
+    (state: RootState) => state.notificationState.legalFactDownloadUrl
+  );
+  const detailTableRows: Array<{ id: number; label: string; value: ReactNode }> = [
+    { id: 1, label: t('Data'), value: <Box fontWeight={600}>{notification.sentAt}</Box> },
+    { id: 2, label: t('Termini di pagamento'), value: t(`Entro il `) },
+    {
+      id: 3,
+      label: t('Destinatario'),
+      value: <Box fontWeight={600}>{notification.recipients[0]?.taxId}</Box>,
+    },
+    {
+      id: 4,
+      label: t('Cognome Nome'),
+      value: <Box fontWeight={600}>{notification.recipients[0]?.denomination}</Box>,
+    },
+    {
+      id: 6,
+      label: t('Codice IUN annullato'),
+      value: <Box fontWeight={600}>{notification.cancelledIun}</Box>,
+    },
+    { id: 7, label: t('Codice IUN'), value: <Box fontWeight={600}>{notification.iun}</Box> },
+    { id: 8, label: t('Gruppi'), value: '' },
+  ];
+  const documentDowloadHandler = (documentIndex: number) => {
+    void dispatch(getReceivedNotificationDocument({ iun: notification.iun, documentIndex }));
+  };
+  const legalFactDownloadHandler = (legalFact: LegalFactId) => {
+    void dispatch(getReceivedNotificationLegalfact({ iun: notification.iun, legalFact }));
+  };
+
+  const dowloadDocument = (url: string) => {
+    /* eslint-disable functional/immutable-data */
+    const link = document.createElement('a');
+    link.href = url;
+    link.target = '_blank';
+    link.rel = 'noreferrer';
+    link.click();
+    /* eslint-enable functional/immutable-data */
+  };
 
   useEffect(() => {
     if (id) {
       void dispatch(getReceivedNotification(id));
     }
   }, []);
+
+  useEffect(() => {
+    if (documentDownloadUrl) {
+      dowloadDocument(documentDownloadUrl);
+    }
+  }, [documentDownloadUrl]);
+
+  useEffect(() => {
+    if (legalFactDownloadUrl) {
+      dowloadDocument(legalFactDownloadUrl);
+    }
+  }, [legalFactDownloadUrl]);
 
   useEffect(() => () => void dispatch(resetState()), []);
 
@@ -57,9 +116,13 @@ const NotificationDetail = () => {
           </Breadcrumbs>
           <Box sx={{ padding: '20px 0 0 0' }}>
             <TitleBox variantTitle="h4" title={notification.subject}></TitleBox>
-            <DetailTable notification={notification} />
+            <NotificationDetailTable rows={detailTableRows} />
             <Paper sx={{ padding: '24px', marginBottom: '20px' }} className="paperContainer">
-              <DetailDocuments notification={notification} />
+              <NotificationDetailDocuments
+                title={t('Atti Allegati')}
+                documents={notification.documents}
+                clickHandler={documentDowloadHandler}
+              />
             </Paper>
             <Button sx={{ margin: '10px 0' }} variant="outlined" onClick={() => navigate(-1)}>
               {t('button.indietro', { ns: 'common' })}
@@ -68,7 +131,13 @@ const NotificationDetail = () => {
         </Grid>
         <Grid item xs={5}>
           <Box sx={{ backgroundColor: 'white', height: '100%', padding: '24px' }}>
-            <DetailTimeline notification={notification} />
+            <NotificationDetailTimeline
+              timeline={notification.timeline}
+              statusHistory={notification.notificationStatusHistory}
+              title={t('Stato della notifica')}
+              legalFactLabel={t('Attestato opponibile a Terzi')}
+              clickHandler={legalFactDownloadHandler}
+            />
           </Box>
         </Grid>
       </Grid>
