+++ packages/pn-personafisica-webapp/src/utils/status.utility.ts
@@ -1,23 +1,3 @@
-import { getNotificationStatusLabelAndColor, NotificationStatus } from '@pagopa-pn/pn-commons';
-import {
-  NotificationDetailTimeline,
-  NotificationStatusHistory,
-} from '@pagopa-pn/pn-commons';
-
-export function getNotificationStatusLabelAndColorFromTimelineCategory(
-  timelineStep: NotificationDetailTimeline,
-  notificationStatusHistory: Array<NotificationStatusHistory>
-): {
-  color: 'warning' | 'error' | 'success' | 'info' | 'default' | 'primary' | 'secondary' | undefined;
-  label: string;
-  tooltip: string;
-} {
-  const notificationStep = notificationStatusHistory.find((n) =>
-    n.relatedTimelineElements.includes(timelineStep.elementId)
-  );
-  return getNotificationStatusLabelAndColor(notificationStep?.status as NotificationStatus);
-}
-
 export enum DelegationStatus {
   ACTIVE = 'Active',
   PENDING = 'Pending',
